+++ client/src/components/SpaceDetails.md
@@ -1,0 +1,10 @@
+SpaceDetails example:
+
+```js
+<SpaceDetails
+    plant={{
+        image_url:
+            'https://bs.floristic.org/image/o/63073d2fbf45b90701279405ecc2eec0272906ed',
+    }}
+/>
+```
