+++ client/src/components/SpaceOverview.md
@@ -1,0 +1,5 @@
+SpaceOverview example:
+
+```js
+<SpaceOverview spaceTitles={['Vegetables', 'Fruits']} />
+```
